requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
decorator==4.4.2
distro==1.9.0
filelock==3.18.0
Flask==3.1.0
fsspec==2025.3.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
openai==1.78.0
openai-whisper==20240930
packaging==25.0
pillow==10.4.0
proglog==0.1.12
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
pytube==15.0.0
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
yt-dlp==2025.4.30
